+++ crates/biome_json_syntax/src/file_source.rs
@@ -1,13 +1,21 @@
 use biome_rowan::FileSourceError;
 use std::path::Path;
 
+// TODO: Jsonc is not well-defined, so some files may only support comments
+// but no trailing commas. We should reconsider whether to use Jsonc as a variant,
+// or just use something like: JsonWithComments, JsonWithCommentsAndTrailingCommas
+//
+// Currently, the "variant" key and other properties are making things duplicated
+// and error-prone.
+
 #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
 #[derive(
     Debug, Clone, Default, Copy, Eq, PartialEq, Hash, serde::Serialize, serde::Deserialize,
 )]
 pub struct JsonFileSource {
     variant: JsonVariant,
-    allow_trailing_comma: bool,
+    allow_trailing_commas: bool,
+    allow_comments: bool,
 }
 
 #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
@@ -24,28 +32,43 @@ impl JsonFileSource {
     pub fn json() -> Self {
         Self {
             variant: JsonVariant::Standard,
-            allow_trailing_comma: false,
+            allow_trailing_commas: false,
+            allow_comments: false,
         }
     }
 
     pub fn jsonc() -> Self {
         Self {
             variant: JsonVariant::Jsonc,
-            allow_trailing_comma: false,
+            allow_trailing_commas: true,
+            allow_comments: true,
         }
     }
 
-    pub fn with_trailing_comma(mut self, option_value: bool) -> Self {
-        self.allow_trailing_comma = option_value;
+    pub fn with_trailing_commas(mut self, option_value: bool) -> Self {
+        self.allow_trailing_commas = option_value;
         self
     }
 
-    pub fn set_allow_trailing_comma(&mut self, option_value: bool) {
-        self.allow_trailing_comma = option_value;
+    pub fn set_allow_trailing_commas(&mut self, option_value: bool) {
+        self.allow_trailing_commas = option_value;
     }
 
-    pub fn allows_trailing_comma(&self) -> bool {
-        self.allow_trailing_comma
+    pub fn get_allow_trailing_commas(&self) -> bool {
+        self.allow_trailing_commas
+    }
+
+    pub fn with_comments(mut self, option_value: bool) -> Self {
+        self.allow_comments = option_value;
+        self
+    }
+
+    pub fn set_allow_comments(&mut self, option_value: bool) {
+        self.allow_comments = option_value;
+    }
+
+    pub fn get_allow_comments(&self) -> bool {
+        self.allow_comments
     }
 
     pub const fn is_jsonc(&self) -> bool {